Move rabbitmq tests to correct category
