+++ rabbitmq/tests/test_integration.py
@@ -11,6 +11,7 @@ import pytest
 
 from datadog_checks.rabbitmq import RabbitMQ
 
+from . import common, metrics
 from .common import (
     CONFIG,
     CONFIG_DEFAULT_VHOSTS,
@@ -219,3 +220,29 @@ def test_connections(aggregator, check):
         aggregator.assert_metric('rabbitmq.connections', tags=['rabbitmq_vhost:test2'], value=0, count=1)
         aggregator.assert_metric('rabbitmq.connections', count=2)
         aggregator.assert_metric('rabbitmq.connections.state', tags=['rabbitmq_conn_state:running'], value=0, count=0)
+
+
+def test_nodes(aggregator, check):
+
+    # default, node metrics are collected
+    check = RabbitMQ('rabbitmq', {}, instances=[common.CONFIG])
+    check.check(common.CONFIG)
+
+    for m in metrics.COMMON_METRICS:
+        aggregator.assert_metric(m, count=1)
+
+    aggregator.reset()
+
+
+def test_disable_nodes(aggregator, check):
+
+    # node metrics collection disabled in config, node metrics should not appear
+    check = RabbitMQ('rabbitmq', {}, instances=[common.CONFIG_NO_NODES])
+    check.check(common.CONFIG_NO_NODES)
+
+    for m in metrics.COMMON_METRICS:
+        aggregator.assert_metric(m, count=0)
+
+    # check to ensure other metrics are being collected
+    for m in metrics.Q_METRICS:
+        aggregator.assert_metric(m, count=1)
