+++ rabbitmq/tests/test_unit.py
@@ -157,32 +157,6 @@ def test_config(check, test_case, extra_config, expected_http_kwargs):
         r.get.assert_called_with('http://localhost:15672/api/connections', **http_wargs)
 
 
-def test_nodes(aggregator, check):
-
-    # default, node metrics are collected
-    check = RabbitMQ('rabbitmq', {}, instances=[common.CONFIG])
-    check.check(common.CONFIG)
-
-    for m in metrics.COMMON_METRICS:
-        aggregator.assert_metric(m, count=1)
-
-    aggregator.reset()
-
-
-def test_disable_nodes(aggregator, check):
-
-    # node metrics collection disabled in config, node metrics should not appear
-    check = RabbitMQ('rabbitmq', {}, instances=[common.CONFIG_NO_NODES])
-    check.check(common.CONFIG_NO_NODES)
-
-    for m in metrics.COMMON_METRICS:
-        aggregator.assert_metric(m, count=0)
-
-    # check to ensure other metrics are being collected
-    for m in metrics.Q_METRICS:
-        aggregator.assert_metric(m, count=1)
-
-
 def test_queues_regexes_exclude_with_negative_lookahead(aggregator, dd_run_check):
     """Based on a support case where a customer was confused why their regular expression didn't work.
